Code Clean 2
Made GameObjectData extend GameObject and also removed the need to copy items from list to list.. That probably took a lot of time. Disadvantage is there is way too much getters and setters in GameObject class but oh well.. Let me know if this code fix is a bad idea.